Move all commands into subgroups
